[Map] Now using Submit and Cancel.
[PlayerActionInputs] Removed most of Map and Stamp action maps.
