--- Snowy-Path/Assets/Scripts/Map/MapTool.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 557db6b441536ed4a86bf8929f36a8c8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
